Refactor: Feed폴더 components/pages안으로 옮김 #2
